+++ cli/Cargo.toml
@@ -11,7 +11,6 @@ readme = "README.md"
 anyhow = "1.0"
 ostree-ext = { path = "../lib" }
 clap = "2.33.3"
-indicatif = "0.15.0"
 structopt = "0.3.21"
 ostree = { version = "0.11.0", features = ["v2021_2"] }
 libc = "0.2.92"
