+++ lib/Cargo.toml
@@ -16,6 +16,7 @@ cjson = "0.1.1"
 flate2 = { features = ["zlib"], default_features = false, version = "1.0.20" }
 fn-error-context = "0.1.1"
 futures = "0.3.13"
+indicatif = "0.15.0"
 gio = "0.9.1"
 glib = "0.10.3"
 glib-sys = "0.10.1"
@@ -31,6 +32,7 @@ ostree = { features = ["v2021_2"], version = "0.11.0" }
 ostree-sys = "0.7.2"
 serde = { features = ["derive"], version = "1.0.125" }
 serde_json = "1.0.64"
+structopt = "0.3.21"
 tar = "0.4.33"
 tempfile = "3.2.0"
 tokio = { features = ["full"], version = "1" }
